dlp2/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
python-dotenv==1.0.0
bcrypt==4.1.2
Pillow==10.0.1
python-magic==0.4.27
pyotp==2.9.0
qrcode==7.4.2
